Fix SonarCloud reliability bugs on new code
Resolves the legitimate Java reliability issues flagged by the SonarCloud
quality gate (Reliability Rating D on new code):

- S2142 (16): re-interrupt the current thread when catching
  InterruptedException across Daemon, ServiceContext, HibernateContextDAO,
  S3StorageService, Hl7InArchivesMigrateThread, WebDaemon, TestInstallUtil,
  UpdateFilter and InitializationFilter.
- S899 (12): act on the boolean returned by File.delete()/createNewFile()
  (log a warning on failure) in ModuleClassLoader, ModuleFactory, Listener,
  AbstractSnapshotTuner, TestInstallUtil and InitializationFilter.
- S2119 (5): reuse a single Random/SecureRandom instance instead of creating
  one per call; use SecureRandom where the value is a password
  (MigrationHelper, InitializationFilter) and a shared Random elsewhere
  (OpenmrsUtil).
- S2201 (6): replace the unused Collection.size() lazy-init trick with the
  self-documenting Hibernate.initialize(...) in UserContext and
  HibernateContextDAO.
- S1143 (2): remove the throw from the finally block in ORUR01Handler and the
  return from the finally block in DatabaseDetective (the latter could wrongly
  report a populated database as empty when closing the connection failed).
- S2390: break the Result -> EmptyResult static-init cycle with a lazy holder.
- S2153 (2), S2184 (2), S4973, S2674, S2886 (2), S2259 (2): assorted boxing,
  integer-overflow, String comparison, unchecked-read, missing-synchronization
  and possible-NPE fixes.

S1872 hits are left as false positives (they compare Class objects / test
class names that are not on the production classpath, so instanceof cannot be
used). The S2276 sleep-in-synchronized in UpdateFilter is a deliberate
anti-brute-force delay and the plsql BLOCKER lives in a historical migration
file; these need a SonarCloud "won't fix" rather than a code change.

+++ api/src/main/java/org/openmrs/hl7/impl/HL7ServiceImpl.java
@@ -1113,7 +1113,7 @@ public class HL7ServiceImpl extends BaseOpenmrsService implements HL7Service, Re
 			}
 
 			//resolve the appropriate month folder
-			File monthDir = new File(yearDir, df.format(calendar.get(Calendar.MONTH) + 1));
+			File monthDir = new File(yearDir, df.format((long) calendar.get(Calendar.MONTH) + 1));
 			if (!monthDir.isDirectory()) {
 				monthDir.mkdirs();
 			}
